requirements.txt
# Web Framework & Real-time
Flask==3.0.0
Flask-SocketIO==5.3.6
eventlet==0.33.3

# Database & Authentication
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2

# AI & Machine Learning
openai==1.12.0
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
numpy==1.26.4

# Scraping & Data Fetching
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11

# Task Scheduling & Utilities
APScheduler==3.10.4
python-dotenv==1.0.1